creating m3u
Today I made writem3u.py which generates the m3u file, and it opened
successfully. The next step is to integrate this into the main file,
but I have had difficulty
of trying to store a user input as a value that I can use throughout
code. I considered using globals for testing purposes, but moved
against it.